# [IMP] core: Improve performance of import