--- a/odoo/orm/fields_textual.py
+++ b/odoo/orm/fields_textual.py
@@ -201,7 +201,7 @@ class BaseString(Field[str | typing.Literal[False]]):
             return
 
         # flush dirty None values
-        dirty_records = records & cache.get_dirty_records(records, self)
+        dirty_records = cache.filtered_dirty_records(records, self)
         if any(v is None for v in cache.get_values(dirty_records, self)):
             dirty_records.flush_recordset([self.name])
 
@@ -216,7 +216,7 @@ class BaseString(Field[str | typing.Literal[False]]):
         # model translation
         if not callable(self.translate):
             # invalidate clean fields because them may contain fallback value
-            clean_records = records - cache.get_dirty_records(records, self)
+            clean_records = cache.filtered_clean_records(records, self)
             clean_records.invalidate_recordset([self.name])
             cache.update(records, self, itertools.repeat(cache_value), dirty=True)
             if lang != 'en_US' and not records.env['res.lang']._get_data(code='en_US'):