The goal of these commit is to improve the performance of large csv file loading. In the same time, it also improves the performance of any imports (especially the big one).

### [IMP] core: Improve import performance

db_id_for() is heavily used when we import or load a csv file. It translates a lot of strings (unused at the end) strings with
the old _() API, which adds extra cost to find the right environment in the stack.

Use self.env._() in all the ir_fields.py file and change the API of db_id_for() to avoid returning field_type because it is never used.

Improvements:
To load l10n_us_hr_payroll/data/res.city.csv, the convert_to_record() took 15 sec, and now it takes 4.5 sec during an upgrade of l10n_us_hr_payroll modules.

### [FIX] core: avoid quadratic complexity of _extract_records()

islice() with a start index actually iterates and ignores elements until the start index is reached (see https://docs.python.org/3/library/itertools.html#itertools.islice). In this case, it leads to a quadratic behavior because we loop once on the data and a second time inside the islice() to reach at least the start index.

We make our own iterator that starts directly at the right place, since we know that the data is a list.

Performance improvements:
To load l10n_us_hr_payroll/data/res.city.csv, the convert_to_record() took 4.5 sec, and now it takes 2 sec during an upgrade of the l10n_us_hr_payroll module.

### [IMP] core: improve performance BaseString.write() for translated field.

BaseString.write() for a translated field becomes slower the more dirty data we have in the cache for that field. This can drastically slow down the import of records that already exist in the DB.

Improvements:
To reload l10n_us_hr_payroll/data/res.city.csv, the _load_records_write() took 20 sec, and now it takes 1.9 sec during an upgrade of the l10n_us_hr_payroll module.

### Global improvements

With the profiling running (with biased of the profiling)

- During installation of l10n_us_hr_payroll, load `l10n_us_hr_payroll/data/res.city.csv` 19.04 sec -> 5.18 sec:
Profiling before: [profile_before_install.json](https://github.com/user-attachments/files/18785414/profile_before_install.json)
Profiling after: [profile_after_install.json](https://github.com/user-attachments/files/18784762/profile_after.json)

- During upgrade of l10n_us_hr_payroll, load `l10n_us_hr_payroll/data/res.city.csv` 45.94 sec -> 5.19 sec:
Profiling before: [profile_before_upgrade.json](https://github.com/user-attachments/files/18785415/profile_before_upgrade.json)
Profiling after: [profile_after_upgrade.json](https://github.com/user-attachments/files/18784784/profile_2.json)