@robodoo 2025-02-12
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/197524.png)](https://mergebot.odoo.com/odoo/odoo/pull/197524)